+++ app/src/main/java/com/smashing/app/core/designsystem/component/sport/SportSelector.kt
@@ -1,3 +1,5 @@
+@file:JvmName("SportSelectorKt")
+
 package com.smashing.app.core.designsystem.component.sport
 
 import androidx.compose.foundation.background
@@ -7,10 +9,6 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
@@ -20,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import com.smashing.app.R.drawable.ic_badminton
 import com.smashing.app.R.drawable.ic_pingpong
 import com.smashing.app.R.drawable.ic_tennis
-import com.smashing.app.R.string.sign_up_sport_title
 import com.smashing.app.R.string.sign_up_sport_subtitle
+import com.smashing.app.R.string.sign_up_sport_title
 import com.smashing.app.core.common.type.SportType
 import com.smashing.app.core.designsystem.component.chip.SmashingChip
 import com.smashing.app.core.designsystem.style.ChipStyle
@@ -33,37 +31,39 @@ import kotlinx.collections.immutable.persistentListOf
 
 
 @Composable
-fun SportSelector (
+fun SportSelector(
     items: ImmutableList<SportType>,
     selectedSport: SportType?,
     onSportSelected: (SportType) -> Unit,
     modifier: Modifier = Modifier,
+    title: String = stringResource(sign_up_sport_title),
+    subTitle: String = stringResource(sign_up_sport_subtitle),
 ) {
-    Column (
+    Column(
         modifier = modifier,
-    ){
+    ) {
         SignUpTitle(
-            title = stringResource(sign_up_sport_title),
-            subTitle = stringResource(sign_up_sport_subtitle),
+            title = title,
+            subTitle = subTitle,
         )
 
-        Row (
+        Row(
             modifier = Modifier
                 .fillMaxWidth(),
-        ){
+        ) {
             items.forEachIndexed { index, item ->
                 SmashingChip(
                     text = item.sportName,
-                    style = if(selectedSport == item) ChipStyle.ACTIVE else ChipStyle.INACTIVE,
+                    style = if (selectedSport == item) ChipStyle.ACTIVE else ChipStyle.INACTIVE,
                     onClick = { onSportSelected(item) },
-                    icon = when(item) {
+                    icon = when (item) {
                         SportType.BADMINTON -> ImageVector.vectorResource(ic_badminton)
                         SportType.PING_PONG -> ImageVector.vectorResource(ic_pingpong)
                         SportType.TENNIS -> ImageVector.vectorResource(ic_tennis)
                     },
                 )
 
-                if(index != items.lastIndex){
+                if (index != items.lastIndex) {
                     Spacer(modifier = Modifier.width(10.dp))
                 }
             }
